+++ plugins/psslurm/src/psslurm/psslurmstep.h
@@ -2,7 +2,7 @@
  * ParaStation
  *
  * Copyright (C) 2017-2021 ParTec Cluster Competence Center GmbH, Munich
- * Copyright (C) 2021-2025 ParTec AG, Munich
+ * Copyright (C) 2021-2026 ParTec AG, Munich
  *
  * This file may be distributed under the terms of the Q Public License
  * as defined in the file LICENSE.QPL included in the packaging of this
@@ -419,7 +419,7 @@ Step_t *__Step_findByEnv(env_t env, uint32_t *jobidOut, uint32_t *stepidOut,
  *
  * Returns the number of steps
  */
-int Step_count(void);
+size_t Step_count(void);
 
 /**
  * @brief Send a signal to all steps of a job
